LayoutTests/platform/ios-sim-deprecated/media/video-source-type-params-expected.txt

EVENT(loadstart)
EXPECTED (stripExtension(relativeURL(video.currentSrc)) == 'content/test'), OBSERVED '' FAIL
END OF TEST
